+++ asilib/examples/fisheye_image_arc.py
@@ -10,7 +10,7 @@ time = datetime(2017, 9, 15, 2, 34, 0)
 
 # A bright auroral arc that was analyzed by Imajo et al., 2021 "Active
 # auroral arc powered by accelerated electrons from very high altitudes"
-image_time, image, ax, im = asilib.plot_image(
+image_time, image, ax, im = asilib.plot_fisheye(
     asi_array_code, location_code, time, color_norm='log', force_download=False
 )
 plt.colorbar(im)