+++ asilib/plot/plot_fisheye.py
@@ -1,6 +1,5 @@
-import pathlib
-from datetime import datetime, timedelta
-from typing import List, Union, Optional, Sequence, Tuple
+from datetime import datetime
+from typing import List, Union, Tuple
 import warnings
 
 import matplotlib
@@ -56,6 +55,36 @@ def plot_image(
     color_norm: str = 'log',
     azel_contours: bool = False,
 ) -> Tuple[datetime, plt.Axes, matplotlib.image.AxesImage]:
+    warnings.warn('asilib.plot_image is renamed to asilib.plot_fisheye and is now deprecated.')
+
+    return plot_fisheye(
+        asi_array_code,
+        location_code,
+        time,
+        force_download=force_download,
+        time_thresh_s=time_thresh_s,
+        ax=ax,
+        label=label,
+        color_map=color_map,
+        color_bounds=color_bounds,
+        color_norm=color_norm,
+        azel_contours=azel_contours,
+    )
+
+
+def plot_fisheye(
+    asi_array_code: str,
+    location_code: str,
+    time: utils._time_type,
+    force_download: bool = False,
+    time_thresh_s: float = 3,
+    ax: plt.Axes = None,
+    label: bool = True,
+    color_map: str = 'auto',
+    color_bounds: List[float] = None,
+    color_norm: str = 'log',
+    azel_contours: bool = False,
+) -> Tuple[datetime, plt.Axes, matplotlib.image.AxesImage]:
     """
     Plots one fisheye image, oriented with North on the top, and East on the right of the image.
 
@@ -71,8 +100,8 @@ def plot_image(
     time_range: list of datetime.datetimes or stings
         Defined the duration of data to download. Must be of length 2.
     force_download: bool
-        If True, download the file even if it already exists. Useful if a prior 
-        data download was incomplete. 
+        If True, download the file even if it already exists. Useful if a prior
+        data download was incomplete.
     time_thresh_s: float
         The maximum allowable time difference between ``time`` and an ASI time stamp.
     ax: plt.Axes
@@ -87,7 +116,8 @@ def plot_image(
     color_bounds: List[float]
         The lower and upper values of the color scale. If None, will
         automatically set it to low=1st_quartile and
-        high=min(3rd_quartile, 10*1st_quartile)
+        high=min(3rd_quartile, 10*1st_quartile). This range works well
+        for most cases.
     color_norm: str
         Sets the 'lin' linear or 'log' logarithmic color normalization.
     azel_contours: bool
@@ -126,7 +156,7 @@ def plot_image(
     | # A bright auroral arc that was analyzed by Imajo et al., 2021 "Active
     | # auroral arc powered by accelerated electrons from very high altitudes"
     | time = datetime(2017, 9, 15, 2, 34, 0)
-    | image_time, ax, im = asilib.plot_image('THEMIS', 'RANK', time,
+    | image_time, ax, im = asilib.plot_fisheye('THEMIS', 'RANK', time,
     |     color_norm='log', force_download=False)
     |
     | plt.colorbar(im)