+++ asilib/tests/test_download_themis.py
@@ -3,11 +3,10 @@ import requests
 import pathlib
 from datetime import datetime
 
-from asilib.io import download_themis
 import asilib
 
 """
-Unit tests to check that the functions in download_themis.py are working correctly.
+Unit tests to check that the functions in download.py are working correctly.
 Run with "python3 test_download_themis.py -v" for the verbose output.
 """
 
@@ -17,8 +16,10 @@ class TestDownloadThemis(unittest.TestCase):
         """Set up a few variables."""
         self.day = datetime(2016, 10, 29, 4)
         self.location_code = 'Gill'
+        self.asi_array_code = 'THEMIs'
+
         self.url = (
-            download_themis.IMG_BASE_URL
+            'http://themis.ssl.berkeley.edu/data/themis/thg/l1/asi/'
             + f'{self.location_code.lower()}/{self.day.year}/{str(self.day.month).zfill(2)}/'
         )
         return
@@ -33,7 +34,7 @@ class TestDownloadThemis(unittest.TestCase):
         self.assertNotEqual(status_code // 100, 5)
         return
 
-    def test_download_themis_img_one_file(self):
+    def test_download_img(self):
         """
         Test the full THEMIS data downloader and download an hour file
         clg_l1_rgf_luck_2020080104_v01.cdf to ./themis/.
@@ -41,9 +42,11 @@ class TestDownloadThemis(unittest.TestCase):
         temp_image_dir = pathlib.Path(asilib.config['ASI_DATA_DIR'], 'themis')
         temp_image_path = temp_image_dir / 'thg_l1_asf_gill_2016102904_v01.cdf'
 
-        download_themis.download_themis_img(self.location_code, self.day, force_download=True)
+        save_path = asilib.download_image(self.asi_array_code, self.location_code, time=self.day, 
+            force_download=True)
 
-        self.assertTrue(temp_image_path.is_file())
+        assert temp_image_path == save_path[0]
+        assert temp_image_path.is_file()
         return